fix: hide scheduled-only temporary terminals
